requirements.txt
PyGithub
hgapi
requests
html5lib
mercurial
